the structure of the slices is divided into folders
