+++ src/redux/slices/mainSlice/actions/mainAction.ts
@@ -1,0 +1,12 @@
+import { createAsyncThunk } from '@reduxjs/toolkit';
+import axios from 'axios';
+import { ToDo } from '../mainSlice';
+
+export const fetchData = createAsyncThunk('data/fetchData', async () => {
+  try {
+    const { data } = await axios.get<ToDo[]>('http://localhost:5500/');
+    return data;
+  } catch (error) {
+    console.log((error as Error).message);
+  }
+});
