+++ src/redux/slices/mainSlice/reducers/mainReducer.ts
@@ -1,0 +1,3 @@
+export const setItemsReducer = (state: any, action: any) => {
+  state.data = action.payload;
+};
