+++ src/redux/slices/mainSlice/selectors/mainSelector.ts
@@ -1,0 +1,3 @@
+import { RootState } from "../../../store";
+
+export const selectToDoData = (state: RootState) => state.data;
