+++ src/redux/slices/mainSlice/types/mainTypes.ts
@@ -1,0 +1,5 @@
+export const enum Status {
+  LOADING = 'loading',
+  SUCCESS = 'success',
+  ERROR = 'error',
+}
